Move `uncommittedHistory` to `SignalDispatchingEntity`
Move the uncommitted-events machinery from `Aggregate` up to
`SignalDispatchingEntity`: the `uncommittedHistory` field and the
`recordEvents` / `getUncommittedEvents` / `hasUncommittedEvents` /
`uncommittedHistory()` / `commitEvents` methods (now `@Internal public`), plus
the `UncommittedHistory` and `UncommittedEvents` classes, moved to
`io.spine.server.entity` and made `@Internal public`.
